+++ gazebo/physics/ode/ODEUniversalJoint.cc
@@ -137,7 +137,7 @@ void ODEUniversalJoint::SetVelocity(int _index, double _angle)
 //////////////////////////////////////////////////
 void ODEUniversalJoint::SetForce(int _index, double _torque)
 {
-  if (_index < 0 || _index >= this->GetAngleCount())
+  if (_index < 0 || static_cast<unsigned int>(_index) >= this->GetAngleCount())
   {
     gzerr << "Calling ODEUniversalJoint::SetForce with an index ["
           << _index << "] out of range\n";
